requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI and ML Libraries
langchain==0.0.350
langchain-community==0.0.1
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.1
accelerate==0.25.0

# Database and Storage
supabase==2.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Frontend and UI
streamlit==1.29.0
gradio==4.7.1

# WhatsApp Integration
twilio==8.11.0
requests==2.31.0

# Utilities
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.2
httpx==0.25.2
aiofiles==23.2.1

# Security and Encryption
cryptography==41.0.8
bcrypt==4.1.2
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Deployment
gunicorn==21.2.0
docker==6.1.3
